Move file conversion code to module in paper_gpu package
