+++ control_software/src/paper_gpu/__init__.py
@@ -11,3 +11,4 @@ except DistributionNotFound:
     pass
 
 from . import bda
+from . import file_conversion
